requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
pydantic>=2.7
python-dotenv>=1.0
email-validator>=2.2
